requirements.txt
fastapi
uvicorn
pillow
python-multipart
piexif>=1.1.3
cryptography>=41.0.0
pydantic>=2.0
